Changing xpm images by Nerd icons in Xmobar
The Nerd icons only work when used within modules and with one of the
additional fonts. That's why Hack Nerd has been added as an additional
font as well as the main font and that's also why the Haskell icon has
been maintained. The extra spacing added to the templates is to prevent
the Nerd font characters to overlap with the text afterwards.