Changed pulse audio shortcut to be a preferences shortcut
Now, when pressing mod+shift+p, the preferences menu is displayed. At
the moment, it only manages audio by executing pulse mixer and vpn by
executing the VPN script.

Another change has been stop relying in the repos/dotfiles path and
assume that the scripts will be in the path. In a later commit the
.profile file will be moved to the repo folder and managed using
Ansible.